pyproject.toml
[project]
name = "medassist-ai"
version = "0.1.0"
description = "MedAssist-AI backend + frontend scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "python-multipart",
  "httpx",
  "pyyaml",
  "streamlit>=1.50.0",
  "requests>=2.32.5",
  "openai>=2.6.0",
  "tiktoken>=0.12.0",
  "torchxrayvision>=1.4.0",
  "scikit-image>=0.25.2",
  "pdfplumber>=0.11.7",
  "torch>=2.9.0",
  "torchvision>=0.24.0",
  "pillow>=11.3.0",
  "numpy>=2.3.4",
  "pydicom>=3.0.1",
  "pymupdf>=1.26.5",
  "pypdf>=6.1.2",
  "pyjwt>=2.9.0",
]

[tool.uv]
# no env vars here
